move source range data into Sources component and make it state
